@@ -7,26 +7,24 @@ import { Reader } from 'protobufjs'
 class ProtobufMessageStream<T> extends Transform {
   private buffer = Buffer.alloc(0)
 
-  constructor(private readonly decodeDelimited: (reader: Reader|Uint8Array) => T) {
-    super({objectMode: true})
+  constructor(
+    private readonly decodeDelimited: (reader: Reader | Uint8Array) => T
+  ) {
+    super({ objectMode: true })
   }
 
-  _transform(chunk: any, encoding: string, callback: TransformCallback) {
+  public _transform(chunk: any, encoding: string, callback: TransformCallback) {
     this.buffer = Buffer.concat([this.buffer, chunk])
-    const reader = Reader.create(this.buffer)
 
-    try {
-      const {len, pos} = reader
-      while (pos < len) {
+    while (true) {
+      try {
+        const reader = Reader.create(this.buffer)
         const message = this.decodeDelimited(reader)
-        if(!message) {
-          return callback(new Error(`No message returned. len=${len}, pos=${pos}`))
-        }
-        this.buffer = this.buffer.slice(reader.pos)
         this.push(message)
+        this.buffer = this.buffer.slice(reader.pos)
+      } catch (err) {
+        break
       }
-    } catch (err) {
-      // wait for more
     }
     callback()
   }
