@@ -1,8 +1,5 @@
 import ProtobufMessageStream from './ProtobufMessageStream'
-import { io } from "./cucumber-messages"
+import { io } from './cucumber-messages'
 import messages = io.cucumber.messages
 
-export {
-  messages,
-  ProtobufMessageStream
-}
+export { messages, ProtobufMessageStream }
